getPitchRange
